+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { db } from 'firebase/clientApp'
 import type { InferGetStaticPropsType, NextPage } from 'next'
 import { Container } from 'src/components/share/Container'
-import { Theme } from 'src/components/Theme'
+import { ThemeList } from 'src/components/share/List'
 
 type Props = InferGetStaticPropsType<typeof getStaticProps>
 
@@ -12,7 +12,7 @@ const Home: NextPage<Props> = ({ articles }) => {
         <ul>
           {articles.map((item) => (
             <li className="" key={item.id}>
-              <Theme id={item.id} theme={item.theme} name={item.name} date={item.date} time={item.time} />
+              <ThemeList id={item.id} theme={item.theme} name={item.name} date={item.date} time={item.time} />
             </li>
           ))}
         </ul>
